install: Move mntdevice dir into baseline
We only need our own mount points (including devtmpfs) in the baseline path.
